Add clarity on installing dependencies
As a newcomer to Elm, i couldn't start elm-pages to work after following the steps. I later figured out that many had the same starting issue (like https://discourse.nixos.org/t/installing-elm-pages-and-other-packages-that-depend-on-elm/8525/3 and https://issueexplorer.com/issue/dillonkearns/elm-pages-starter/29) and i though it would have been of great benefit if this missing piece is included.